Plot labeling curves
Source: R/visualization.R

Visualizes the labeling fits for peptides or proteins. Selects a random subset from the provided data and plots a maximum of plot_number of plots.

tor_plot_labeling_curves(data, grouping = protein, time_col = "hours",
  plot_number = 6, random_seed = random())

Arguments

data
data with label rate (and error), growth rate (and error)

grouping
by protein or peptide for visualization ??

time_col
the name of the column that holds the time, assumes "hours" by default.

plot_number
designate the maximum number of curves you want to plot. Use plot_number = NULL to plot all - be careful with this option, the plot might take a long time to render with a larger number of proteins or peptides.

random_seed
specify this to get reproducible selection of "random" plots